Fix: 타입 에러 수정 및 Flutter SDK 업데이트
- upgrade_replay_service.dart 타입 캐스팅 에러 수정
- Flutter SDK 3.35.2 업데이트 (업그레이드 진행)
- 패키지 의존성 재설치
- Chrome 브라우저에서 정상 빌드 확인

+++ lib/services/upgrade_replay_service.dart
@@ -342,7 +342,7 @@ class UpgradeReplayService {
       .where((e) => e.value > 0)
       .map((e) => {'category': e.key, 'count': e.value})
       .toList()
-      ..sort((a, b) => b['count'].compareTo(a['count']));
+      ..sort((a, b) => (b['count'] as int).compareTo(a['count'] as int));
   }
   
   // 8. 다음 학습 단계 제안
